+++ webpack.config.js
@@ -72,7 +72,7 @@ module.exports = (env, argv) => {
           { from: 'icon.png', to: 'icon.png', noErrorOnMissing: true },
           { from: 'preview.png', to: 'preview.png', noErrorOnMissing: true },
           { from: 'README.md', to: 'README.md', noErrorOnMissing: true },
-          { from: 'README_zh_CN.md', to: 'README_zh_CN.md', noErrorOnMissing: true },
+          { from: 'README_zh-CN.md', to: 'README_zh-CN.md', noErrorOnMissing: true },
         ],
       }),
       ...(isProd ? [